BIM: consolidate wall geometry value objects
Keep wall baseline, path, and resolved section data in one geometry module with strict immutable value objects.

Rename the endpoint helper to reflect its wall-specific responsibility, update ArchWall and direct geometry tests to use the final module boundaries, and register only the resulting modules in the BIM build.
