+++ src/Mod/BIM/Arch.py
@@ -2151,163 +2151,17 @@ def makeWindow(
 
 
 def is_debasable(wall):
-    """Determines if an Arch Wall can be cleanly converted to a baseless state.
+    """Return whether an Arch wall can be converted to a baseless wall."""
+    import ArchWallEndpoint
 
-    This function checks if a given wall is a valid candidate for a parametric
-    "debasing" operation, where its dependency on a Base object is removed and
-    it becomes driven by its own Length and Placement properties.
-
-    Parameters
-    ----------
-    wall : FreeCAD.DocumentObject
-        The Arch Wall object to check.
-
-    Returns
-    -------
-    bool
-        ``True`` if the wall is a valid candidate for debasing, otherwise ``False``.
-
-    Notes
-    -----
-    A wall is considered debasable if its ``Base`` object's final shape consists
-    of exactly one single, straight edge. This check is generic and works for
-    any base object that provides a valid ``.Shape`` property, including
-    ``Draft.Line`` and ``Sketcher::SketchObject`` objects.
-    """
-    import Part
-    import Draft
-
-    # Ensure the object is actually a wall
-    if Draft.getType(wall) != "Wall":
-        return False
-
-    # Check for a valid Base object with a geometric Shape
-    if not hasattr(wall, "Base") or not wall.Base:
-        return False
-    if not hasattr(wall.Base, "Shape") or wall.Base.Shape.isNull():
-        return False
-
-    base_shape = wall.Base.Shape
-
-    # The core condition: the final shape must contain exactly one edge.
-    # This correctly handles Sketches with multiple lines or construction geometry.
-    if len(base_shape.Edges) != 1:
-        return False
-
-    # The single edge must be a straight line.
-    edge = base_shape.Edges[0]
-    if not isinstance(edge.Curve, (Part.Line, Part.LineSegment)):
-        return False
-
-    # If all checks pass, the wall is debasable.
-    return True
+    return ArchWallEndpoint.is_debasable(wall)
 
 
 def debaseWall(wall):
-    """
-    Converts a line-based Arch Wall to be parametrically driven by its own
-    properties (Length, Width, Height) and Placement, removing its dependency
-    on a Base object.
+    """Convert a line-based Arch wall to a baseless wall in place."""
+    import ArchWallEndpoint
 
-    This operation preserves the wall's exact size and global position.
-    It is only supported for walls based on a single, straight line.
-
-    Parameters
-    ----------
-    wall : FreeCAD.DocumentObject
-        The Arch Wall object to debase.
-
-    Returns
-    -------
-    bool
-        True on success, False otherwise.
-    """
-    import FreeCAD
-
-    if not is_debasable(wall):
-        FreeCAD.Console.PrintWarning(f"Wall '{wall.Label}' is not eligible for debasing.\n")
-        return False
-
-    doc = wall.Document
-
-    try:
-        # Record the current global placements of all children that move with the host.
-        # ArchComponent.onChanged will attempt to shift them when the wall's placement is updated
-        # below.
-        children = wall.Proxy.getMovableChildren(wall)
-        child_placements = {child: child.Placement.copy() for child in children}
-
-        # --- Calculation of the final placement ---
-        base_obj = wall.Base
-        base_edge = base_obj.Shape.Edges[0]
-
-        # Step 1: Get global coordinates of the baseline's endpoints.
-        # For Draft objects, Vertex coordinates are already in the global system. For Sketches,
-        # they are local, but ArchWall's internal logic transforms them. The most reliable
-        # way to get the final global baseline is to use the vertices of the base object's
-        # final shape, which are always in global coordinates for these object types.
-        p1_global = base_edge.Vertexes[0].Point
-        p2_global = base_edge.Vertexes[1].Point
-
-        # Step 2: Determine the extrusion normal vector.
-        normal = wall.Normal
-        if normal.Length == 0:
-            normal = base_obj.Placement.Rotation.multVec(FreeCAD.Vector(0, 0, 1))
-            if normal.Length == 0:
-                normal = FreeCAD.Vector(0, 0, 1)
-
-        # Step 3: Calculate the final orientation from the geometric axes.
-        # - The local Z-axis is the extrusion direction (height).
-        # - The local X-axis is along the baseline (length).
-        # - The local Y-axis is perpendicular to both, pointing "Right" to remain
-        #   consistent with the wall's internal creation logic (X x Z = Y).
-        z_axis = normal.normalize()
-        x_axis = (p2_global - p1_global).normalize()
-        y_axis = x_axis.cross(z_axis).normalize()
-        final_rotation = FreeCAD.Rotation(x_axis, y_axis, z_axis)
-
-        # Step 4: Calculate the final position (the wall's volumetric center).
-        # The new placement's Base must be the global coordinate of the final wall's center.
-        centerline_position = (p1_global + p2_global) * 0.5
-
-        # The new placement's Base is the center of the baseline. The alignment is handled by the
-        # geometry generation itself, not by shifting the placement.
-        final_position = centerline_position
-        final_placement = FreeCAD.Placement(final_position, final_rotation)
-
-        # Store properties before unlinking
-        height = wall.Height.Value
-        length = wall.Length.Value
-        width = wall.Width.Value
-
-        # 1. Apply the final placement first.
-        wall.Placement = final_placement
-
-        # Restore original placements to counteract the shift from onChanged.
-        # This keeps all hosted elements stationary in world space.
-        for child, original_placement in child_placements.items():
-            child.Placement = original_placement
-
-        # 2. Now, remove the base. The recompute triggered by this change
-        #    will already have the correct placement to work with.
-        wall.Base = None
-
-        # 3. Clear internal caches and set final properties.
-        if hasattr(wall.Proxy, "connectEdges"):
-            wall.Proxy.connectEdges = []
-
-        wall.Height = height
-        wall.Length = length
-        wall.Width = width
-
-        # 4. Add an explicit recompute to ensure the final state is settled.
-        doc.recompute()
-
-    except Exception as e:
-        FreeCAD.Console.PrintError(f"Error debasing wall '{wall.Label}': {e}\n")
-        return False
-
-    return True
+    return ArchWallEndpoint.debaseWall(wall)
 
 
 def _initializeArchObject(
