+++ src/Mod/BIM/ArchWallGeometry.py
@@ -18,11 +18,15 @@
 #                                                                              #
 ################################################################################
 
-"""Value objects and operations for an already-resolved finite wall baseline.
+"""Pure value objects and operations for resolved wall geometry.
 
 ``WallBaseline`` and ``WallPath`` are deliberately independent of wall
 proxies and FreeCAD wall properties.  Callers provide global straight
-geometry; this module validates it and exposes path-to-path operations only.
+geometry; this module validates it and exposes path-to-path operations.
+
+``WallSection`` and ``WallSectionLayer`` represent an already-resolved wall
+profile.  They expose geometric queries only and never resolve FreeCAD
+properties or wall proxies.
 """
 
 from dataclasses import dataclass
@@ -178,3 +182,64 @@ def find_path_intersection(path_a, path_b):
         return None, None, None
     point = intersections[0]
     return point, path_a.nearest_end_name(point), path_b.nearest_end_name(point)
+
+
+@dataclass(frozen=True)
+class WallSectionLayer:
+    """One resolved material layer in local lateral coordinates.
+
+    ``raw_thickness`` retains the source sign.  Positive layers are visible;
+    negative layers are construction-only cursor steps.  ``y_min`` and
+    ``y_max`` describe the layer's resolved lateral interval.
+    """
+
+    raw_thickness: float
+    y_min: float
+    y_max: float
+
+    @property
+    def visible(self):
+        return self.raw_thickness > 0
+
+
+@dataclass(frozen=True)
+class WallSection:
+    """Immutable resolved wall section consumed by relation geometry.
+
+    The layer tuple is already ordered and positioned by ``ArchWall``.  This
+    value object exposes only geometric queries and never resolves FreeCAD
+    properties or wall proxies.
+    """
+
+    layers: tuple
+
+    @property
+    def visible_layers(self):
+        return tuple(layer for layer in self.layers if layer.visible)
+
+    @property
+    def y_min(self):
+        return min((layer.y_min for layer in self.visible_layers), default=0.0)
+
+    @property
+    def y_max(self):
+        return max((layer.y_max for layer in self.visible_layers), default=0.0)
+
+    def offset_towards(self, lateral_direction, world_direction):
+        """Return the signed lateral offset to the requested visible face.
+
+        A zero world direction has no preferred face, so it selects the
+        ``y_min`` face deterministically.  Callers that require a directional
+        result should resolve or reject that ambiguity before calling here.
+        """
+        if lateral_direction is None or world_direction is None:
+            return None
+        if lateral_direction.Length <= 1e-9:
+            return None
+        if not self.visible_layers:
+            return None
+        if world_direction.Length <= 1e-9:
+            return -self.y_min
+        if lateral_direction.dot(world_direction) >= 0:
+            return -self.y_min
+        return -self.y_max
